Revert "teams2"
This reverts commit 182d1229fc142064897ff7d650dec8103fd2b309.

+++ src/main/java/com/clicker/Clicker/entities/Team.java
@@ -1,6 +1,5 @@
 package com.clicker.Clicker.entities;
 
-import com.clicker.Clicker.entities.items.Item;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
@@ -8,14 +7,14 @@ import java.util.List;
 
 @Entity
 @Table(name="Teams")
-public class Team{
+public class Team {
 
     @Id
     @Column(name = "team_name")
     private String team_name;
 
     @Column(name="click_count")
-    private long click_count = 0;
+    private int click_count = 0;
 
     @OneToOne
     @JoinColumn(name="admin")
@@ -35,11 +34,11 @@ public class Team{
         this.team_name = team_name;
     }
 
-    public long getClick_count() {
+    public int getClick_count() {
         return click_count;
     }
 
-    public void setClick_count(long click_count) {
+    public void setClick_count(int click_count) {
         this.click_count = click_count;
     }
